Fixes crash due to inconsistent data types

+++ piwigo/Piwigo Data Wrappers/PiwigoTagData.h
@@ -12,7 +12,7 @@
 
 @property (nonatomic, assign) NSInteger tagId;
 @property (nonatomic, strong) NSString *tagName;
-@property (nonatomic, assign) NSInteger numberOfImagesUnderTag;
+@property (nonatomic, assign) long long numberOfImagesUnderTag;
 @property (nonatomic, strong) NSDate *lastModified;
 
 @end
